face-recognition-login/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
face-recognition
opencv-python
Pillow
numpy